Modularizing third party auth.
