+++ src/account-settings/actions.js
@@ -10,8 +10,6 @@ export const OPEN_FORM = 'OPEN_FORM';
 export const CLOSE_FORM = 'CLOSE_FORM';
 export const UPDATE_DRAFT = 'UPDATE_DRAFT';
 export const RESET_DRAFTS = 'RESET_DRAFTS';
-export const DISCONNECT_AUTH = new AsyncActionType('ACCOUNT_SETTINGS', 'DISCONNECT_AUTH');
-
 
 // FETCH SETTINGS ACTIONS
 
@@ -113,24 +111,4 @@ export const fetchTimeZones = country => ({
 export const fetchTimeZonesSuccess = timeZones => ({
   type: FETCH_TIME_ZONES.SUCCESS,
   payload: { timeZones },
-});
-
-
-// DISCONNECT AUTH ACTIONS
-
-export const disconnectAuth = (url, providerId) => ({
-  type: DISCONNECT_AUTH.BASE, payload: { url, providerId },
-});
-export const disconnectAuthBegin = providerId => ({
-  type: DISCONNECT_AUTH.BEGIN, payload: { providerId },
-});
-export const disconnectAuthSuccess = (providerId, thirdPartyAuthProviders) => ({
-  type: DISCONNECT_AUTH.SUCCESS,
-  payload: { providerId, thirdPartyAuthProviders },
-});
-export const disconnectAuthFailure = providerId => ({
-  type: DISCONNECT_AUTH.FAILURE, payload: { providerId },
-});
-export const disconnectAuthReset = providerId => ({
-  type: DISCONNECT_AUTH.RESET, payload: { providerId },
 });
