+++ src/account-settings/third-party-auth/index.js
@@ -1,0 +1,5 @@
+export { default } from './ThirdPartyAuth';
+export { default as reducer } from './data/reducers';
+export { default as saga } from './data/sagas';
+export { configureService, getThirdPartyAuthProviders } from './data/service';
+export { DISCONNECT_AUTH } from './data/actions';
